refactor(toast): co-locate toast helper inside components/Toast
- Move the compat shim from src/utils/toast.ts into src/components/Toast/toast.ts so the helper lives next to the component, provider, types, and stories it belongs with.
- Drop the dead pre-sonner index.ts (the old middle-API implementation, no longer imported anywhere).
